Revert "refactor(#29): 1주차 pr 리뷰 반영"
This reverts commit 477baf9d195f45f3c14b93c72aca471e295a38e0.
