+++ be/issue_tracker/src/main/java/team1/issue_tracker/Issue/IssueListRes.java
@@ -2,7 +2,6 @@ package team1.issue_tracker.Issue;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import team1.issue_tracker.comment.Comment;
 import team1.issue_tracker.label.Label;
 
 import java.util.List;
@@ -13,7 +12,7 @@ import team1.issue_tracker.milestone.Milestone;
 public class IssueListRes {
     private long id;
     private String title;
-    private Comment comment;
+    private String comment;
     private List<Label> labels;
     private Milestone milestone;
 }
