+++ be/issue_tracker/src/main/java/team1/issue_tracker/Issue/IssueService.java
@@ -1,9 +1,8 @@
 package team1.issue_tracker.Issue;
 
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import team1.issue_tracker.comment.Comment;
-import team1.issue_tracker.comment.CommentRepository;
 import team1.issue_tracker.label.IssueLabel;
 import team1.issue_tracker.label.IssueLabelRepository;
 import team1.issue_tracker.label.Label;
@@ -17,18 +16,15 @@ import team1.issue_tracker.milestone.MilestoneRepository;
 public class IssueService {
 
     private final IssueRepository issueRepository;
-    private final CommentRepository commentRepository;
     private final LabelRepository labelRepository;
     private final IssueLabelRepository issueLabelRepository;
     private final MilestoneRepository milestoneRepository;
 
     @Autowired
-    public IssueService(IssueRepository issueRepository, CommentRepository commentRepository,
-                        LabelRepository labelRepository,
+    public IssueService(IssueRepository issueRepository, LabelRepository labelRepository,
                         IssueLabelRepository issueLabelRepository,
                         MilestoneRepository milestoneRepository) {
         this.issueRepository = issueRepository;
-        this.commentRepository = commentRepository;
         this.labelRepository = labelRepository;
         this.issueLabelRepository = issueLabelRepository;
         this.milestoneRepository = milestoneRepository;
@@ -37,8 +33,7 @@ public class IssueService {
     public List<IssueListRes> getList() {
         List<Issue> issueList = (List<Issue>) issueRepository.findAll();
 
-        return issueList.stream().map(issue -> new IssueListRes(issue.getId(), issue.getTitle(),
-                commentAtIssue(issue.getId()),
+        return issueList.stream().map(issue -> new IssueListRes(issue.getId(), issue.getTitle(), issue.getComment(),
                 labelsAtIssue(issue.getId()), getMilestone(issue))).toList();
     }
 
@@ -50,17 +45,9 @@ public class IssueService {
     }
 
     private List<Label> labelsAtIssue(Long issueId) {
-        List<IssueLabel> labels = issueLabelRepository.findAllByIssueId(issueId);
+        List<IssueLabel> byIssueId = issueLabelRepository.findAllByIssueId(issueId);
 
-        return labels.stream().map(label -> labelRepository.findById(label.getLabelId()).get()
+        return byIssueId.stream().map(issueLabel -> labelRepository.findById(issueLabel.getLabelId()).get()
         ).toList();
     }
-
-    private Comment commentAtIssue(Long issueId) {
-        return commentRepository.findFirstBy(issueId);
-    }
-
-//    private List<Comment> commentsAtIssue(Long issueId) {
-//        return commentRepository.findAllByIssueId(issueId);
-//    }
 }
