--- be/issue_tracker/src/main/java/team1/issue_tracker/comment/CommentRepository.java
@@ -1,11 +1,0 @@
-package team1.issue_tracker.comment;
-
-import java.util.List;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CommentRepository extends CrudRepository<Comment, Long> {
-    List<Comment> findAllByIssueId(Long issueId);
-    Comment findFirstBy(Long issueId);
-}
