+++ be/issue_tracker/src/main/java/team1/issue_tracker/temp/User.java
@@ -1,20 +1,20 @@
-package team1.issue_tracker.user;
+package team1.issue_tracker.temp;
 
-import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.relational.core.mapping.Column;
 import org.springframework.data.relational.core.mapping.Table;
 
+import java.sql.Timestamp;
+
 @Table("USERS")
-@Getter
-@AllArgsConstructor
 public class User {
     @Id
     private String id;
     private String password;
     private String name;
+    @Column("created_at")
+    private Timestamp createdAt;
+    @Column("profile_img")
     private String profileImg;
     private UserType type;
-    private LocalDateTime createdAt;
 }
