+++ be/issue_tracker/src/main/java/team1/issue_tracker/temp/UserType.java
@@ -1,4 +1,4 @@
-package team1.issue_tracker.user;
+package team1.issue_tracker.temp;
 
 public enum UserType {
     NORMAL, GITHUB, APPLE
