+++ be/issue_tracker/src/main/resources/sql/ddl.sql
@@ -1,84 +1,65 @@
-drop table if exists ISSUE_HAS_ASSIGNEE, ISSUE_HAS_LABEL, USER_LIKES_COMMENT, COMMENT_FILE, COMMENT, ASSIGNEE, ISSUE, LABEL, MILESTONE, USERS;
+drop table if exists ISSUE_LABEL, ISSUE_MILESTONE, LABEL, MILESTONE, COMMENT, ISSUE, USERS;
 
 create table USERS
 (
     id          varchar(255) primary key,
     password    varchar(255),
-    name        varchar(255) not null,
+    name        varchar(255)                       not null,
+    created_at  timestamp                          not null default current_timestamp,
     profile_img varchar(255),
-    login_type  varchar(50)  not null default 'NORMAL',
-    created_at  timestamp             default current_timestamp
+    type        enum ('normal', 'github', 'apple') not null default 'normal'
 );
 
 create table MILESTONE
 (
     id          bigint auto_increment primary key,
-    name        varchar(255) unique not null,
-    description text,
-    deadline    timestamp,
-    created_at  timestamp default current_timestamp
-);
-
-create table LABEL
-(
-    id          bigint auto_increment primary key,
-    name        varchar(255) unique not null,
-    description text,
-    color       varchar(7)          not null default '#ffffff',
-    created_at  timestamp                    default current_timestamp
+    name        varchar(255) not null,
+    description text         not null,
+    created_at  timestamp    not null default current_timestamp,
+    deadline    timestamp
 );
 
 create table ISSUE
 (
     id               bigint auto_increment primary key,
-    user_id          varchar(255) not null,
-    milestone_id     bigint,
-    title            varchar(255) not null,
-    status           varchar(50)  not null default 'OPEN',
-    last_modified_at timestamp             default current_timestamp,
-    created_at       timestamp             default current_timestamp,
-    foreign key (user_id) references USERS (id),
-    foreign key (milestone_id) references MILESTONE (id)
+    user_id          varchar(255)           not null,
+    title            varchar(255)           not null,
+    created_at       timestamp              not null default current_timestamp,
+    last_modified_at timestamp              not null default current_timestamp,
+    status           enum ('open', 'close') not null default 'open',
+    foreign key (user_id) references USERS (id)
 );
 
-create table ASSIGNEE
+create table LABEL
 (
-    id       bigint auto_increment primary key,
-    issue_id bigint,
-    user_id  varchar(255),
-    foreign key (issue_id) references ISSUE (id),
-    foreign key (user_id) references USERS (id)
+    id          bigint auto_increment primary key,
+    name        varchar(255) not null,
+    description text         not null,
+    created_at  timestamp    not null default current_timestamp,
+    color       varchar(7)   not null default '#ffffff'
 );
 
 create table COMMENT
 (
     id               bigint auto_increment primary key,
-    issue_id         bigint,
-    user_id          varchar(255),
-    last_modified_at timestamp default current_timestamp,
-    created_at       timestamp default current_timestamp,
-    foreign key (issue_id) references ISSUE (id),
-    foreign key (user_id) references USERS (id)
-);
-
-create table COMMENT_FILE
-(
-    id         bigint auto_increment primary key,
-    comment_id bigint,
-    file_url   varchar(255) not null,
-    foreign key (comment_id) references COMMENT (id)
+    user_id          varchar(255) not null,
+    issue_id         bigint       not null,
+    created_at       timestamp    not null default current_timestamp,
+    last_modified_at timestamp    not null default current_timestamp,
+    foreign key (user_id) references USERS (id),
+    foreign key (issue_id) references ISSUE (id)
 );
 
 create table USER_LIKES_COMMENT
 (
-    user_id    varchar(255),
-    comment_id bigint,
+    user_id    varchar(255) not null,
+    comment_id bigint       not null,
     primary key (user_id, comment_id),
     foreign key (user_id) references USERS (id),
     foreign key (comment_id) references COMMENT (id)
 );
 
-create table ISSUE_HAS_LABEL
+create table ISSUE_LABEL
 (
     id       bigint auto_increment primary key,
     issue_id bigint,
@@ -88,12 +69,20 @@ create table ISSUE_HAS_LABEL
     unique (issue_id, label_id)
 );
 
-create table ISSUE_HAS_ASSIGNEE
+create table ASSIGNEE
 (
-    id          bigint auto_increment primary key,
-    issue_id    bigint,
-    assignee_id bigint,
+    id       bigint auto_increment primary key,
+    issue_id bigint,
+    user_id  varchar(255) not null,
     foreign key (issue_id) references ISSUE (id),
-    foreign key (assignee_id) references ASSIGNEE (id),
-    unique (issue_id, assignee_id)
+    foreign key (user_id) references LABEL (id),
+    unique (issue_id, user_id)
+);
+
+create table COMMENT_FILE
+(
+    id         bigint auto_increment primary key,
+    comment_id bigint,
+    file_url   varchar(255) not null,
+    foreign key (comment_id) references COMMENT (id)
 );
